K64F: Explicitly set the .text section alignment
If you don't specify the alignment of a section explicitly, that section
inherits it from the element with the highest alignment inside that
section.

Since the uVisor code is in the .text section, and it *must* end up at a
known location, we set the alignment of .text to 4 so that its start
address ends up at a deterministic location.